[types] Move test to tests file.
