+++ cmd/starcoin/src/account/accept_token_cmd.rs
@@ -3,7 +3,7 @@
 
 use crate::cli_state::CliState;
 use crate::StarcoinOpt;
-use anyhow::{bail, format_err, Result};
+use anyhow::{format_err, Result};
 use scmd::{CommandAction, ExecContext};
 use starcoin_crypto::hash::HashValue;
 use starcoin_executor::DEFAULT_EXPIRATION_TIME;
