+++ cmd/starcoin/src/account/execute_builtin_script_cmd.rs
@@ -3,7 +3,7 @@
 
 use crate::cli_state::CliState;
 use crate::StarcoinOpt;
-use anyhow::{bail, Result};
+use anyhow::{format_err, Result};
 use scmd::{CommandAction, ExecContext};
 use starcoin_crypto::hash::HashValue;
 use starcoin_rpc_client::RemoteStateReader;
